Remove obsolete test files for tool call middleware
- Deleted several test files related to tool call middleware, including `index.prompt-template.test.ts`, `tool-call-middleware.convert-prompt.test.ts`, `tool-call-middleware.errors.test.ts`, `tool-call-middleware.merge-users.test.ts`, and `tool-call-middleware.positive.test.ts`.
- This cleanup improves the overall structure and maintainability of the test suite by removing outdated and redundant tests.

These changes enhance the clarity and focus of the testing framework in the parser package.
