+++ packages/parser/tests/index.test.ts
@@ -104,185 +104,6 @@ describe("index exports", () => {
     });
   });
 
-  describe("non-stream assistant->user merge formatting with object input", () => {
-    it("gemma: formats assistant tool-call (object input) and tool result into user text", async () => {
-      const mw = gemmaToolMiddleware;
-
-      const out = await mw.transformParams!({
-        params: {
-          prompt: [
-            { role: "user", content: [{ type: "text", text: "q" }] },
-            {
-              role: "assistant",
-              content: [
-                {
-                  type: "tool-call",
-                  toolCallId: "tc1",
-                  toolName: "get_weather",
-                  // simulate provider giving parsed object input
-                  input: JSON.stringify({ city: "Seoul" }),
-                } as any,
-              ],
-            },
-            {
-              role: "tool",
-              content: [
-                {
-                  type: "tool-result",
-                  toolName: "get_weather",
-                  toolCallId: "tc1",
-                  output: { ok: true },
-                },
-              ],
-            },
-          ],
-          tools: [
-            {
-              type: "function",
-              name: "get_weather",
-              description: "",
-              inputSchema: { type: "object" },
-            },
-          ],
-        },
-      } as any);
-
-      // last message is the tool result
-      console.debug(out.prompt[out.prompt.length - 1]);
-
-      const assistantMsg = out.prompt.find((m: any) => m.role === "assistant")!;
-      const assistantText = (assistantMsg.content as any[])
-        .map((c: any) => (c.type === "text" ? c.text : ""))
-        .join("");
-      expect(assistantText).toMatch(/tool_call/);
-
-      const userMsgs = out.prompt.filter((m: any) => m.role === "user");
-      const userCombined = userMsgs
-        .map((u: any) =>
-          u.content.map((c: any) => (c.type === "text" ? c.text : "")).join("")
-        )
-        .join("\n");
-
-      expect(userCombined).toMatch(/tool_response/);
-    });
-
-    it("hermes: formats assistant tool-call (object input) and tool result into user text", async () => {
-      const mw = hermesToolMiddleware;
-      const out = await mw.transformParams!({
-        params: {
-          prompt: [
-            { role: "user", content: [{ type: "text", text: "q" }] },
-            {
-              role: "assistant",
-              content: [
-                {
-                  type: "tool-call",
-                  toolCallId: "tc1",
-                  toolName: "get_weather",
-                  input: JSON.stringify({ city: "Seoul" }),
-                } as any,
-              ],
-            },
-            {
-              role: "tool",
-              content: [
-                {
-                  type: "tool-result",
-                  toolName: "get_weather",
-                  toolCallId: "tc1",
-                  output: { ok: true },
-                },
-              ],
-            },
-          ],
-          tools: [
-            {
-              type: "function",
-              name: "get_weather",
-              description: "",
-              inputSchema: { type: "object" },
-            },
-          ],
-        },
-      } as any);
-
-      // last message is the tool result
-      console.debug(out.prompt[out.prompt.length - 1]);
-
-      const assistantMsg = out.prompt.find((m: any) => m.role === "assistant")!;
-      const assistantText = (assistantMsg.content as any[])
-        .map((c: any) => (c.type === "text" ? c.text : ""))
-        .join("");
-      expect(assistantText).toMatch(/<tool_call>/);
-
-      const userMsgs = out.prompt.filter((m: any) => m.role === "user");
-      const userCombined = userMsgs
-        .map((u: any) =>
-          u.content.map((c: any) => (c.type === "text" ? c.text : "")).join("")
-        )
-        .join("\n");
-      expect(userCombined).toMatch(/<tool_response>/);
-    });
-
-    it("xml: formats assistant tool-call (object input) and tool result into user text", async () => {
-      const mw = xmlToolMiddleware;
-      const out = await mw.transformParams!({
-        params: {
-          prompt: [
-            { role: "user", content: [{ type: "text", text: "q" }] },
-            {
-              role: "assistant",
-              content: [
-                {
-                  type: "tool-call",
-                  toolCallId: "tc1",
-                  toolName: "get_weather",
-                  input: JSON.stringify({ city: "Seoul" }),
-                } as any,
-              ],
-            },
-            {
-              role: "tool",
-              content: [
-                {
-                  type: "tool-result",
-                  toolName: "get_weather",
-                  toolCallId: "tc1",
-                  output: { ok: true },
-                },
-              ],
-            },
-          ],
-          tools: [
-            {
-              type: "function",
-              name: "get_weather",
-              description: "",
-              inputSchema: { type: "object" },
-            },
-          ],
-        },
-      } as any);
-
-      // last message is the tool result
-      console.debug(out.prompt[out.prompt.length - 1]);
-
-      const assistantMsg = out.prompt.find((m: any) => m.role === "assistant")!;
-      const assistantText = (assistantMsg.content as any[])
-        .map((c: any) => (c.type === "text" ? c.text : ""))
-        .join("");
-      expect(assistantText).toMatch(/<get_weather>/);
-
-      const userMsgs = out.prompt.filter((m: any) => m.role === "user");
-      const userCombined = userMsgs
-        .map((u: any) =>
-          u.content.map((c: any) => (c.type === "text" ? c.text : "")).join("")
-        )
-        .join("\n");
-      expect(userCombined).toMatch(/<tool_response>/);
-    });
-  });
-
   describe("createToolMiddleware export", () => {
     it("should be exported and callable", () => {
       expect(createToolMiddleware).toBeDefined();
