+++ packages/parser/tests/tool-call-middleware.branches.test.ts
@@ -47,72 +47,6 @@ describe("createToolMiddleware branches", () => {
     });
   });
 
-  it("transformParams throws on toolChoice type none", async () => {
-    const mw = createToolMiddleware({
-      protocol: jsonMixProtocol,
-      toolSystemPromptTemplate: () => "",
-    });
-    await expect(
-      mw.transformParams!({
-        params: { prompt: [], tools: [], toolChoice: { type: "none" } },
-      } as any)
-    ).rejects.toThrow(/none/);
-  });
-
-  it("transformParams validates specific tool selection and builds JSON schema", async () => {
-    const mw = createToolMiddleware({
-      protocol: jsonMixProtocol,
-      toolSystemPromptTemplate: () => "",
-    });
-    const tools = [
-      {
-        type: "function",
-        name: "t1",
-        description: "d",
-        inputSchema: { type: "object", properties: { a: { type: "string" } } },
-      },
-    ];
-    const result = await mw.transformParams!({
-      params: {
-        prompt: [],
-        tools,
-        toolChoice: { type: "tool", toolName: "t1" },
-      },
-    } as any);
-    expect(result.responseFormat).toMatchObject({ type: "json", name: "t1" });
-    expect(
-      (result.providerOptions as any).toolCallMiddleware.toolChoice
-    ).toEqual({ type: "tool", toolName: "t1" });
-  });
-
-  it("transformParams required builds if/then/else schema", async () => {
-    const mw = createToolMiddleware({
-      protocol: jsonMixProtocol,
-      toolSystemPromptTemplate: () => "",
-    });
-    const tools = [
-      {
-        type: "function",
-        name: "a",
-        description: "",
-        inputSchema: { type: "object" },
-      },
-      {
-        type: "function",
-        name: "b",
-        description: "",
-        inputSchema: { type: "object" },
-      },
-    ];
-    const result = await mw.transformParams!({
-      params: { prompt: [], tools, toolChoice: { type: "required" } },
-    } as any);
-    expect(result.responseFormat).toMatchObject({ type: "json" });
-    expect(
-      (result.providerOptions as any).toolCallMiddleware.toolChoice
-    ).toEqual({ type: "required" });
-  });
-
   it("wrapStream delegates to toolChoiceStream when toolChoice 'required' is active", async () => {
     const mw = createToolMiddleware({
       protocol: jsonMixProtocol,
